created separate styling file for event card
